refactor: move SectionLabel to component/ui
